refactor: move edge reader
