+++ include/functions/table/hop.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "readers/adj_reader.hpp"
+#include "readers/low_edge_reader.hpp"
 #include "utils/func.hpp"
 
 #include <duckdb/common/named_parameter_map.hpp>
